pyproject.toml
[project]
name = "swimcuttimes"
version = "0.1.0"
description = "Swim cut times tracking for club, high school, and college swimmers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "typer>=0.15",
    "rich>=13.0",
    "httpx>=0.28",
    "anthropic>=0.42",
    "python-dotenv>=1.0",
    "supabase>=2.0",
    "structlog>=25.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]

[tool.hatch.build.targets.wheel]
packages = ["src/swimcuttimes"]
